Application/requirements.txt
# Requirements for the project
streamlit==1.33.0
numpy==1.26.0
pandas==2.2.2
pymongo==4.7.0
torch==2.3.0
torchvision==0.18.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0